streamlit/pyproject.toml
[project]
name = "dashboard"
version = "0"
requires-python = ">=3.13"
dependencies = [
    "altair>=6.0",
    "streamlit>=1.55",
    "plotly>=6.6",
    "google-cloud-bigquery>=3.25",
    "polars>=1.39.3"
]
